requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
Faker==25.1.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
